pyproject.toml
[project]
name = "med_triage_env"
version = "0.1.0"
description = "Real-world Medical Triage environment for OpenEnv training."
authors = [
    { name = "Gemini AI", email = "gemini@example.com" }
]
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "fastmcp",
    "requests",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
